move Doc to overflowdb package (next step: move to odb repo)
