+++ semanticcpg/src/main/scala/io/shiftleft/overflowdb/traversal/Doc.scala
@@ -1,4 +1,4 @@
-package io.shiftleft.semanticcpg
+package io.shiftleft.overflowdb.traversal
 
 import scala.annotation.StaticAnnotation
 import scala.reflect.runtime.universe._
