+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/language/Steps.scala
@@ -5,7 +5,7 @@ import io.shiftleft.codepropertygraph.generated.nodes
 import io.shiftleft.codepropertygraph.generated.nodes._
 import java.util.{List => JList}
 
-import io.shiftleft.semanticcpg.Doc
+import io.shiftleft.overflowdb.traversal.Doc
 import org.apache.tinkerpop.gremlin.process.traversal.Scope
 import org.json4s.CustomSerializer
 import org.json4s.native.Serialization.{write, writePretty}
